requirements.txt
requests
beautifulsoup4
sentence-transformers
numpy
scikit-learn
google-generativeai
python-dotenv